+++ sdk/webpubsub/Azure.Messaging.WebPubSub/src/WebPubSubServiceClient_helpers.cs
@@ -30,7 +30,7 @@ namespace Azure.Messaging.WebPubSub
         internal static byte[] s_group = Encoding.UTF8.GetBytes("webpubsub.group");
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
         /// <param name="expiresAt">UTC time when the token expires.</param>
         /// <param name="userId">User Id.</param>
@@ -40,10 +40,10 @@ namespace Azure.Messaging.WebPubSub
 #pragma warning disable AZC0015 // Unexpected client method return type.
         public virtual Uri GetClientAccessUri(DateTimeOffset expiresAt, string userId, IEnumerable<string> roles, CancellationToken cancellationToken)
 #pragma warning restore AZC0015 // Unexpected client method return type.
-            => GetClientAccessUriInternal(expiresAt, userId, roles, null, async: false, cancellationToken).EnsureCompleted();
+            => GetClientAccessUriInternal(expiresAt, userId, roles, null, ClientEndpointType.Default, async: false, cancellationToken).EnsureCompleted();
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
         /// <param name="expiresAt">UTC time when the token expires.</param>
         /// <param name="userId">User Id.</param>
@@ -52,12 +52,27 @@ namespace Azure.Messaging.WebPubSub
         /// <param name="groups">Groups that the connection with the generated token will join when it connects.</param>
         /// <returns></returns>
 #pragma warning disable AZC0015 // Unexpected client method return type.
-        public virtual Uri GetClientAccessUri(DateTimeOffset expiresAt, string userId = default, IEnumerable<string> roles = default, IEnumerable<string> groups = default, CancellationToken cancellationToken = default)
+        public virtual Uri GetClientAccessUri(DateTimeOffset expiresAt, string userId, IEnumerable<string> roles, IEnumerable<string> groups, CancellationToken cancellationToken)
 #pragma warning restore AZC0015 // Unexpected client method return type.
-            => GetClientAccessUriInternal(expiresAt, userId, roles, groups, async: false, cancellationToken).EnsureCompleted();
+            => GetClientAccessUriInternal(expiresAt, userId, roles, groups, ClientEndpointType.Default, async: false, cancellationToken).EnsureCompleted();
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
+        /// </summary>
+        /// <param name="expiresAt">UTC time when the token expires.</param>
+        /// <param name="userId">User Id.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <param name="roles">Roles that the connection with the generated token will have.</param>
+        /// <param name="groups">Groups that the connection with the generated token will join when it connects.</param>
+        /// <param name="clientEndpointType">The client endpoint type.</param>
+        /// <returns></returns>
+#pragma warning disable AZC0015 // Unexpected client method return type.
+        public virtual Uri GetClientAccessUri(DateTimeOffset expiresAt, string userId = default, IEnumerable<string> roles = default, IEnumerable<string> groups = default, ClientEndpointType clientEndpointType = default, CancellationToken cancellationToken = default)
+#pragma warning restore AZC0015 // Unexpected client method return type.
+            => GetClientAccessUriInternal(expiresAt, userId, roles, groups, clientEndpointType, async: false, cancellationToken).EnsureCompleted();
+
+        /// <summary>
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
         /// <param name="expiresAt">UTC time when the token expires.</param>
         /// <param name="userId">User Id.</param>
@@ -66,11 +81,11 @@ namespace Azure.Messaging.WebPubSub
         /// <returns></returns>
 #pragma warning disable AZC0015 // Unexpected client method return type.
         public virtual async Task<Uri> GetClientAccessUriAsync(DateTimeOffset expiresAt, string userId, IEnumerable<string> roles, CancellationToken cancellationToken)
-            => await GetClientAccessUriInternal(expiresAt, userId, roles, null, async: true, cancellationToken).ConfigureAwait(false);
+            => await GetClientAccessUriInternal(expiresAt, userId, roles, null, ClientEndpointType.Default, async: true, cancellationToken).ConfigureAwait(false);
 #pragma warning restore AZC0015 // Unexpected client method return type.
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
         /// <param name="expiresAt">UTC time when the token expires.</param>
         /// <param name="userId">User Id.</param>
@@ -79,13 +94,29 @@ namespace Azure.Messaging.WebPubSub
         /// <param name="groups">Groups that the connection with the generated token will join when it connects.</param>
         /// <returns></returns>
 #pragma warning disable AZC0015 // Unexpected client method return type.
-        public virtual async Task<Uri> GetClientAccessUriAsync(DateTimeOffset expiresAt, string userId = default, IEnumerable<string> roles = default, IEnumerable<string> groups = default, CancellationToken cancellationToken = default)
-            => await GetClientAccessUriInternal(expiresAt, userId, roles, groups, async: true, cancellationToken).ConfigureAwait(false);
+        public virtual async Task<Uri> GetClientAccessUriAsync(DateTimeOffset expiresAt, string userId, IEnumerable<string> roles, IEnumerable<string> groups, CancellationToken cancellationToken)
+            => await GetClientAccessUriInternal(expiresAt, userId, roles, groups, ClientEndpointType.Default, async: true, cancellationToken).ConfigureAwait(false);
 #pragma warning restore AZC0015 // Unexpected client method return type.
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
+        /// <param name="expiresAt">UTC time when the token expires.</param>
+        /// <param name="clientEndpointType">The client endpoint type.</param>
+        /// <param name="userId">User Id.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <param name="roles">Roles that the connection with the generated token will have.</param>
+        /// <param name="groups">Groups that the connection with the generated token will join when it connects.</param>
+        /// <returns></returns>
+#pragma warning disable AZC0015 // Unexpected client method return type.
+        public virtual async Task<Uri> GetClientAccessUriAsync(DateTimeOffset expiresAt, string userId = default, IEnumerable<string> roles = default, IEnumerable<string> groups = default, ClientEndpointType clientEndpointType = default, CancellationToken cancellationToken = default)
+            => await GetClientAccessUriInternal(expiresAt, userId, roles, groups, clientEndpointType, async: true, cancellationToken).ConfigureAwait(false);
+#pragma warning restore AZC0015 // Unexpected client method return type.
+
+        /// <summary>
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
+        /// </summary>
+        /// <remarks>Not applicable for MQTT clients who connect to "/clients/mqtt/hubs/{hub}"</remarks>
         /// <param name="expiresAfter">Defaults to one hour, if not specified. Must be greater or equal zero.</param>
         /// <param name="userId">User Id.</param>
         /// <param name="roles">Roles that the connection with the generated token will have.</param>
@@ -99,11 +130,11 @@ namespace Azure.Messaging.WebPubSub
             CancellationToken cancellationToken)
 #pragma warning restore AZC0015 // Unexpected client method return type.
         {
-            return GetClientAccessUriInternal(expiresAfter, userId, roles, null, false, cancellationToken).EnsureCompleted();
+            return GetClientAccessUriInternal(expiresAfter, userId, roles, null, ClientEndpointType.Default, false, cancellationToken).EnsureCompleted();
         }
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
         /// <param name="expiresAfter">Defaults to one hour, if not specified. Must be greater or equal zero.</param>
         /// <param name="userId">User Id.</param>
@@ -113,19 +144,43 @@ namespace Azure.Messaging.WebPubSub
         /// <returns></returns>
 #pragma warning disable AZC0015 // Unexpected client method return type.
         public virtual Uri GetClientAccessUri(
+            TimeSpan expiresAfter,
+            string userId,
+            IEnumerable<string> roles,
+            IEnumerable<string> groups,
+            CancellationToken cancellationToken)
+#pragma warning restore AZC0015 // Unexpected client method return type.
+        {
+            return GetClientAccessUriInternal(expiresAfter, userId, roles, groups, ClientEndpointType.Default, false, cancellationToken).EnsureCompleted();
+        }
+
+        /// <summary>
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
+        /// </summary>
+        /// <param name="expiresAfter">Defaults to one hour, if not specified. Must be greater or equal zero.</param>
+        /// <param name="userId">User Id.</param>
+        /// <param name="roles">Roles that the connection with the generated token will have.</param>
+        /// <param name="groups">Groups that the connection with the generated token will join when it connects.</param>
+        /// <param name="clientEndpointType">The client endpoint type.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <returns></returns>
+#pragma warning disable AZC0015 // Unexpected client method return type.
+        public virtual Uri GetClientAccessUri(
             TimeSpan expiresAfter = default,
             string userId = default,
             IEnumerable<string> roles = default,
             IEnumerable<string> groups = default,
+            ClientEndpointType clientEndpointType = default,
             CancellationToken cancellationToken = default)
 #pragma warning restore AZC0015 // Unexpected client method return type.
         {
-            return GetClientAccessUriInternal(expiresAfter, userId, roles, groups, false, cancellationToken).EnsureCompleted();
+            return GetClientAccessUriInternal(expiresAfter, userId, roles, groups, clientEndpointType, false, cancellationToken).EnsureCompleted();
         }
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
+        /// <remarks>Not applicable for MQTT clients who connect to "/clients/mqtt/hubs/{hub}"</remarks>
         /// <param name="expiresAfter">Defaults to one hour, if not specified. Must be greater or equal zero.</param>
         /// <param name="userId">User Id.</param>
         /// <param name="roles">Roles that the connection with the generated token will have.</param>
@@ -139,11 +194,11 @@ namespace Azure.Messaging.WebPubSub
             CancellationToken cancellationToken)
 #pragma warning restore AZC0015 // Unexpected client method return type.
         {
-            return await GetClientAccessUriInternal(expiresAfter, userId, roles, null, true, cancellationToken).ConfigureAwait(false);
+            return await GetClientAccessUriInternal(expiresAfter, userId, roles, null, ClientEndpointType.Default, true, cancellationToken).ConfigureAwait(false);
         }
 
         /// <summary>
-        /// Creates a URI with authentication token.
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
         /// </summary>
         /// <param name="expiresAfter">Defaults to one hour, if not specified. Must be greater or equal zero.</param>
         /// <param name="userId">User Id.</param>
@@ -153,14 +208,37 @@ namespace Azure.Messaging.WebPubSub
         /// <returns></returns>
 #pragma warning disable AZC0015 // Unexpected client method return type.
         public virtual async Task<Uri> GetClientAccessUriAsync(
+            TimeSpan expiresAfter,
+            string userId,
+            IEnumerable<string> roles,
+            IEnumerable<string> groups,
+            CancellationToken cancellationToken)
+#pragma warning restore AZC0015 // Unexpected client method return type.
+        {
+            return await GetClientAccessUriInternal(expiresAfter, userId, roles, groups, ClientEndpointType.Default, true, cancellationToken).ConfigureAwait(false);
+        }
+
+        /// <summary>
+        /// Creates a URI with authentication token for clients who connect to "/clients/hubs/{hub}".
+        /// </summary>
+        /// <param name="expiresAfter">Defaults to one hour, if not specified. Must be greater or equal zero.</param>
+        /// <param name="userId">User Id.</param>
+        /// <param name="roles">Roles that the connection with the generated token will have.</param>
+        /// <param name="groups">Groups that the connection with the generated token will join when it connects.</param>
+        /// <param name="clientEndpointType">The client endpoint type.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <returns></returns>
+#pragma warning disable AZC0015 // Unexpected client method return type.
+        public virtual async Task<Uri> GetClientAccessUriAsync(
             TimeSpan expiresAfter = default,
             string userId = default,
             IEnumerable<string> roles = default,
             IEnumerable<string> groups = default,
+            ClientEndpointType clientEndpointType = default,
             CancellationToken cancellationToken = default)
 #pragma warning restore AZC0015 // Unexpected client method return type.
         {
-            return await GetClientAccessUriInternal(expiresAfter, userId, roles, groups, true, cancellationToken).ConfigureAwait(false);
+            return await GetClientAccessUriInternal(expiresAfter, userId, roles, groups, clientEndpointType, true, cancellationToken).ConfigureAwait(false);
         }
 
         internal static int GetMinutesToExpire(TimeSpan expiresAfter) => Math.Max((int)expiresAfter.TotalMinutes, 1);
@@ -168,85 +246,28 @@ namespace Azure.Messaging.WebPubSub
         internal static int GetMinutesToExpire(DateTimeOffset expiresAt) => Math.Max((int)expiresAt.Subtract(DateTimeOffset.UtcNow).TotalMinutes, 1);
 
         private async Task<Uri> GetClientAccessUriInternal(
-            DateTimeOffset expiresAt,
-            string userId = default,
-            IEnumerable<string> roles = default,
-            IEnumerable<string> groups = default,
-            bool async = true,
-            CancellationToken cancellationToken = default)
+           DateTimeOffset expiresAt,
+           string userId = default,
+           IEnumerable<string> roles = default,
+           IEnumerable<string> groups = default,
+           ClientEndpointType clientEndpointType = default,
+           bool async = true,
+           CancellationToken cancellationToken = default)
         {
-            string token;
-
-            if (_tokenCredential != null)
-            {
-                RequestContext context = new() { CancellationToken = cancellationToken };
-
-                var minutesToExpire = GetMinutesToExpire(expiresAt);
-
-                Response clientTokenResponse = async ?
-                    await GenerateClientTokenImplAsync(userId, roles, minutesToExpire, groups, context).ConfigureAwait(false) :
-                    GenerateClientTokenImpl(userId, roles, minutesToExpire, null, context);
-                using var jsonDocument = JsonDocument.Parse(clientTokenResponse.Content);
-                token = jsonDocument.RootElement.GetProperty(ClientTokenResponseTokenPropertyName).GetString();
-            }
-            else if (_credential != null)
-            {
-                token = GenerateTokenFromAzureKeyCredential(expiresAt, userId, roles, groups);
-            }
-            else
-            {
-                throw new InvalidOperationException($"{nameof(WebPubSubServiceClient)} must be constructed with either a {typeof(TokenCredential)} or {typeof(AzureKeyCredential)} to generate client access URIs.");
-            }
-
-            UriBuilder clientEndpoint = new(Endpoint)
-            {
-                Scheme = Endpoint.Scheme == "http" ? "ws" : "wss"
-            };
-
-            return new Uri($"{clientEndpoint}client/hubs/{_hub}?access_token={token}");
+            var token = await GetClientAccessTokenCore(expiresAt, userId, roles, groups, clientEndpointType, async, cancellationToken: cancellationToken).ConfigureAwait(false);
+            return GetClientAccessUriInternal(token, clientEndpointType);
         }
 
-        private async Task<Uri> GetClientAccessUriInternal(
+        private Task<Uri> GetClientAccessUriInternal(
             TimeSpan expireAfter,
             string userId = default,
             IEnumerable<string> roles = default,
             IEnumerable<string> groups = default,
+            ClientEndpointType clientEndpointType = default,
             bool async = true,
             CancellationToken cancellationToken = default)
         {
-            string token;
-
-            if (_tokenCredential != null)
-            {
-                RequestContext context = new() { CancellationToken = cancellationToken };
-
-                var minutesToExpire = GetMinutesToExpire(expireAfter);
-
-                Response clientTokenResponse = async ?
-                    await GenerateClientTokenImplAsync(userId, roles, minutesToExpire, groups, context).ConfigureAwait(false) :
-                    GenerateClientTokenImpl(userId, roles, minutesToExpire, null, context);
-                using var jsonDocument = JsonDocument.Parse(clientTokenResponse.Content);
-                token = jsonDocument.RootElement.GetProperty(ClientTokenResponseTokenPropertyName).GetString();
-            }
-            else if (_credential != null)
-            {
-                if (expireAfter == default)
-                {
-                    expireAfter = TimeSpan.FromHours(1);
-                }
-                token = GenerateTokenFromAzureKeyCredential(DateTimeOffset.UtcNow.Add(expireAfter), userId, roles, groups);
-            }
-            else
-            {
-                throw new InvalidOperationException($"{nameof(WebPubSubServiceClient)} must be constructed with either a {typeof(TokenCredential)} or {typeof(AzureKeyCredential)} to generate client access URIs.");
-            }
-
-            UriBuilder clientEndpoint = new(Endpoint)
-            {
-                Scheme = Endpoint.Scheme == "http" ? "ws" : "wss"
-            };
-
-            return new Uri($"{clientEndpoint}client/hubs/{_hub}?access_token={token}");
+            return GetClientAccessUriInternal(expireAfter == default ? DateTimeOffset.UtcNow.Add(TimeSpan.FromHours(1)) : DateTimeOffset.UtcNow.Add(expireAfter), userId, roles, groups, clientEndpointType, async, cancellationToken);
         }
 
         /// <summary>
@@ -321,7 +342,7 @@ namespace Azure.Messaging.WebPubSub
             }
         }
 
-        private string GenerateTokenFromAzureKeyCredential(DateTimeOffset expiresAt, string userId = default, IEnumerable<string> roles = default, IEnumerable<string> groups = default)
+        private string GenerateTokenFromAzureKeyCredential(DateTimeOffset expiresAt, ClientEndpointType endpointType, string userId = default, IEnumerable<string> roles = default, IEnumerable<string> groups = default)
         {
             var keyBytes = Encoding.UTF8.GetBytes(_credential.Key);
 
@@ -333,7 +354,7 @@ namespace Azure.Messaging.WebPubSub
             {
                 endpoint += "/";
             }
-            var audience = $"{endpoint}client/hubs/{_hub}";
+            var audience = $"{endpoint}{GetRelativeClientEndpoint(endpointType)}";
 
             if (userId != default)
             {
@@ -353,6 +374,64 @@ namespace Azure.Messaging.WebPubSub
             jwt.AddClaim(JwtBuilder.Aud, audience);
 
             return jwt.BuildString();
+        }
+
+        private string GetRelativeClientEndpoint(ClientEndpointType endpointType) => endpointType switch
+        {
+            ClientEndpointType.Default => $"client/hubs/{_hub}",
+            ClientEndpointType.MQTT => $"clients/mqtt/hubs/{_hub}",
+            _ => throw new ArgumentOutOfRangeException(nameof(endpointType))
+        };
+
+        private async Task<string> GetClientAccessTokenCore(
+             DateTimeOffset expiresAt,
+             string userId,
+             IEnumerable<string> roles,
+             IEnumerable<string> groups,
+             ClientEndpointType clientEndpointType,
+             bool async,
+             CancellationToken cancellationToken = default)
+        {
+            if (clientEndpointType == ClientEndpointType.MQTT && _apiVersionEnum < WebPubSubServiceClientOptions.ServiceVersion.V2024_01_01)
+            {
+                throw new NotSupportedException($"Getting client access token for MQTT is supported since API version {WebPubSubServiceClientOptions.ServiceVersion.V2024_01_01.ToVersionString()}. Current version is {_apiVersion}.");
+            }
+            var clientEndpointString = clientEndpointType switch
+            {
+                ClientEndpointType.Default => "default",
+                ClientEndpointType.MQTT => "mqtt",
+                _ => throw new ArgumentOutOfRangeException(nameof(clientEndpointType))
+            };
+            if (_tokenCredential != null)
+            {
+                RequestContext context = new() { CancellationToken = cancellationToken };
+
+                var minutesToExpire = GetMinutesToExpire(expiresAt);
+
+                Response clientTokenResponse = async ?
+                    await GenerateClientTokenImplAsync(userId, roles, minutesToExpire, groups, clientEndpointString, context).ConfigureAwait(false) :
+                    GenerateClientTokenImpl(userId, roles, minutesToExpire, null, clientEndpointString, context);
+                using var jsonDocument = JsonDocument.Parse(clientTokenResponse.Content);
+                return jsonDocument.RootElement.GetProperty(ClientTokenResponseTokenPropertyName).GetString();
+            }
+            else if (_credential != null)
+            {
+                return GenerateTokenFromAzureKeyCredential(expiresAt, clientEndpointType, userId, roles, groups);
+            }
+            else
+            {
+                throw new InvalidOperationException($"{nameof(WebPubSubServiceClient)} must be constructed with either a {typeof(TokenCredential)} or {typeof(AzureKeyCredential)} to generate client access URIs.");
+            }
+        }
+
+        private Uri GetClientAccessUriInternal(string token, ClientEndpointType endpointType)
+        {
+            UriBuilder clientEndpoint = new(Endpoint)
+            {
+                Scheme = Endpoint.Scheme == "http" ? "ws" : "wss"
+            };
+
+            return new Uri($"{clientEndpoint}{GetRelativeClientEndpoint(endpointType)}?access_token={token}");
         }
     }
 }